(fix)refactor into client/components folder
